requirements.txt
openai>=1.78.0
deepseek>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0